feat: extract template_generation into standalone module with Clean Architecture
Template generation previously lived inside src/planning/domain/generator.rs.
It's now a first-class module with its own Clean Architecture structure.

Created:
- src/template_generation/mod.rs — Module root
- src/template_generation/domain/generator.rs — TemplateGenerator trait,
  ClaudeTemplateGenerator, GeneratorError, RepoContext, GeneratedTemplate,
  InvalidSymbolReference, ClaudeGeneratorConfig (23KB, migrated from planning)
- src/template_generation/domain/event/mod.rs — TemplateGenerationEvent (4 variants)
- src/template_generation/domain/error.rs — Re-exports GeneratorError
- src/template_generation/application/service.rs — TemplateGenerationService trait
- src/template_generation/application/factory.rs — TemplateGenerationFactory trait
- src/template_generation/application/dto/mod.rs — DTOs (GenerateTemplateInput/Output,
  SymbolValidationInput/Output, InvalidSymbolRef, etc.)
- src/template_generation/infrastructure/repository/mod.rs — Re-exports
  GeneratedTemplateRepository from planning
- src/template_generation/interfaces/http/mod.rs — HTTP API contracts
- src/template_generation/tests.rs — 30 unit tests
- src/lib.rs — Registered template_generation module

Updated:
- src/planning/domain/generator.rs — Thin re-export shim (backward compat)
- src/planning/domain/error.rs — Moved From<GeneratorError> for PlanningError here
- .pi/architecture/modules/template-generation.md — Updated file paths

All 889 tests pass (859 existing + 30 new template_generation tests)
