+++ engine/src/template_generation/application/mod.rs
@@ -1,0 +1,8 @@
+//! Application layer for the Template Generation bounded context.
+//!
+//! @canonical .pi/architecture/modules/template-generation.md#application
+//! Issue: issue-contract-freeze
+
+pub mod dto;
+pub mod factory;
+pub mod service;
