+++ engine/src/template_generation/domain/error.rs
@@ -1,0 +1,7 @@
+//! Generator errors - migrated to standalone module.
+//!
+//! @canonical .pi/architecture/modules/template-generation.md#errors
+//! Issue: issue-contract-freeze
+//!
+//! Re-exports from generator.rs for backward compatibility.
+pub use super::generator::GeneratorError;
