+++ engine/src/template_generation/infrastructure/mod.rs
@@ -1,0 +1,6 @@
+//! Infrastructure layer for the Template Generation bounded context.
+//!
+//! @canonical .pi/architecture/modules/template-generation.md#infrastructure
+//! Issue: issue-contract-freeze
+
+pub mod repository;
