+++ engine/src/template_generation/interfaces/mod.rs
@@ -1,0 +1,6 @@
+//! Interfaces layer for the Template Generation bounded context.
+//!
+//! @canonical .pi/architecture/modules/template-generation.md#interfaces
+//! Issue: issue-contract-freeze
+
+pub mod http;
